rtsp2mjpeg/rootfs/opt/rtsp2mjpeg/requirements.txt
rtsp
Pillow
flask[async]